rename useCashu hooks
